+++ src/main/scala/scalismo/faces/image/filter/MorphologicalFilter.scala
@@ -64,157 +64,3 @@ case class MorphologicalFilter[A: ClassTag](structuringElement: PixelImage[Boole
 object MorphologicalFilter {
   def boxElement(size: Int): PixelImage[Boolean] = PixelImage.view(size, size, (x, y) => x >= 0 && x < size && y >= 0 && y < size)
 }
-
-/**
-  * Separable morphological filter (separable structuring element)
-  * @param structuringElement 1d structuring element, must be 1 row or column
-  * @param windowFilter Filtering function
-  */
-case class SeparableMorphologicalFilter[A: ClassTag](structuringElement: PixelImage[Boolean], windowFilter: (Seq[A]) => A) extends ImageFilter[A, A] {
-  assert(structuringElement.width == 1 || structuringElement.height == 1, "Structuring element must be 1D")
-
-  private val width = structuringElement.width
-  private val height = structuringElement.height
-
-  private val rowElement = {
-    if (height == 1)
-      structuringElement
-    else
-      structuringElement.transposed
-  }
-
-  private val colElement = {
-    if (height == 1)
-      structuringElement.transposed
-    else
-      structuringElement
-  }
-
-  private val columnFilter = MorphologicalFilter[A](rowElement, windowFilter)
-  private val rowFilter = MorphologicalFilter[A](colElement, windowFilter)
-
-  override def filter(image: PixelImage[A]): PixelImage[A] = image.domain match {
-    case d: ColumnMajorImageDomain => image.filter(columnFilter).withAccessMode(Repeat()).filter(rowFilter)
-    case d: RowMajorImageDomain => image.filter(rowFilter).withAccessMode(Repeat()).filter(columnFilter)
-  }
-}
-
-object SeparableMorphologicalFilter {
-  def lineElement(size: Int): PixelImage[Boolean] = PixelImage.view(size, 1, (x, y) => x >= 0 && x < size)
-}
-
-object Dilation {
-  /**
-    * dilation filter with box element
-    * @param size side length of box
-    */
-  def box(size: Int): SeparableMorphologicalFilter[Double] = {
-    SeparableMorphologicalFilter(SeparableMorphologicalFilter.lineElement(size), list => list.max)
-  }
-
-  /**
-    * general dilation filter with specified structuring element
-    * @param structuringElement structuring element, see morphological filter
-    */
-  def apply(structuringElement: PixelImage[Boolean]): MorphologicalFilter[Double] = {
-    MorphologicalFilter[Double](structuringElement, values => values.max)
-  }
-}
-
-object Erosion {
-  /**
-    * erosion filter with box element
-    * @param size side length of box
-    */
-  def box(size: Int): SeparableMorphologicalFilter[Double] = {
-    SeparableMorphologicalFilter(SeparableMorphologicalFilter.lineElement(size), list => list.min)
-  }
-
-  /**
-    * general erosion filter with structuring element
-    * @param structuringElement structuring element, see morphological filter
-    */
-  def apply(structuringElement: PixelImage[Boolean]): MorphologicalFilter[Double] = {
-    MorphologicalFilter[Double](structuringElement, values => values.min)
-  }
-}
-
-object Median {
-  // extract median value
-  private def median(values: Seq[Double]): Double = {
-    if(values.length <= 1)
-      values.head
-    else{
-      val sorted = values.sorted
-      (sorted(values.length/2) + sorted(values.length - values.length/2)) / 2
-    }
-  }
-
-  /**
-    * median filter with box as its structuring element
-    * @param size side length of the box
-    */
-  def box(size: Int): MorphologicalFilter[Double] = {
-    MorphologicalFilter(MorphologicalFilter.boxElement(size), median)
-  }
-
-  /**
-    * approximate median filter with box as its structuring element, separable (not identical to box)
-    * @param size side length of the box
-    */
-  def separableBox(size: Int): SeparableMorphologicalFilter[Double] = {
-    SeparableMorphologicalFilter(SeparableMorphologicalFilter.lineElement(size), median)
-  }
-
-  /**
-    * general median filter with structuring element
-    * @param structuringElement structuring element, see morphological filter
-    */
-  def apply(structuringElement: PixelImage[Boolean]): MorphologicalFilter[Double] = {
-    MorphologicalFilter[Double](structuringElement, median)
-  }
-}
-
-object Opening {
-  def box(size: Int): ImageFilter[Double, Double] = {
-    val eroder = Erosion.box(size)
-    val dilator = Dilation.box(size)
-    image => image.filter(eroder).withAccessMode(Repeat()).filter(dilator)
-  }
-
-  def apply(structuringElement: PixelImage[Boolean]): ImageFilter[Double, Double] = {
-    val eroder = Erosion(structuringElement)
-    val dilator = Dilation(structuringElement)
-    image => image.filter(eroder).withAccessMode(Repeat()).filter(dilator)
-  }
-}
-
-object Closing {
-  def box(size: Int): ImageFilter[Double, Double] = {
-    val eroder = Erosion.box(size)
-    val dilator = Dilation.box(size)
-    image => image.filter(dilator).withAccessMode(Repeat()).filter(eroder)
-  }
-
-  def apply(structuringElement: PixelImage[Boolean]): ImageFilter[Double, Double] = {
-    val eroder = Erosion(structuringElement)
-    val dilator = Dilation(structuringElement)
-    image => image.filter(dilator).withAccessMode(Repeat()).filter(eroder)
-  }
-}
-
-object MorphologicalGradient {
-  import PixelImage.implicits._
-
-  def box(size: Int): ImageFilter[Double, Double] = {
-    val eroder = Erosion.box(size)
-    val dilator = Dilation.box(size)
-    image => image.filter(dilator) - image.filter(eroder)
-  }
-
-  def apply(structuringElement: PixelImage[Boolean]): ImageFilter[Double, Double] = {
-    val eroder = Erosion(structuringElement)
-    val dilator = Dilation(structuringElement)
-    image => image.filter(dilator) - image.filter(eroder)
-  }
-}
